workflow errors fixed

+++ anon/backend/src/routes/user.rs
@@ -2,6 +2,7 @@ use axum::{Json, extract::State};
 use serde::Serialize;
 use utoipa::ToSchema;
 
+
 use crate::{
     libs::{db::AppState, error::ApiError},
     middlewares::auth::AuthUser,
@@ -55,7 +56,7 @@ pub async fn me(
     });
 
     Ok(Json(UserMeRes {
-        id: rec.id,
+        id: rec.id as i64,
         wallet: rec.wallet,
         created_at: rec.created_at,
         profile,
